requirements.txt
pillow
pyqtdarktheme
PySide6
